Struct alloy_signer_ledger::LedgerSigner

pub struct LedgerSigner {
    transport: Mutex<Ledger>,
    derivation: DerivationType,
    pub(crate) chain_id: Option<ChainId>,
    pub(crate) address: Address,
}

A Ledger Ethereum signer.

This is a simple wrapper around the Ledger transport.

Note that this wallet only supports asynchronous operations. Calling a non-asynchronous method will always return an error.

Fields
transport: Mutex<Ledger>derivation: DerivationTypechain_id: Option<ChainId>address: Address
Implementations

impl LedgerSigner

pub async fn new( derivation: DerivationType, chain_id: Option<ChainId>, ) -> Result<Self, LedgerError>

Instantiate the application by acquiring a lock on the ledger device.

Examples

use alloy_signer_ledger::{HDPath, LedgerSigner};

let ledger = LedgerSigner::new(HDPath::LedgerLive(0), Some(1)).await?;

pub async fn get_address(&self) -> Result<Address, LedgerError>

Get the account which corresponds to our derivation path

pub async fn get_address_with_path( &self, derivation: &DerivationType, ) -> Result<Address, LedgerError>

Gets the account which corresponds to the provided derivation path

async fn get_address_with_path_transport( transport: &Ledger, derivation: &DerivationType, ) -> Result<Address, LedgerError>

pub async fn version(&self) -> Result<Version, LedgerError>

Returns the semver of the Ethereum ledger app

pub async fn sign_tx_rlp(&self, tx_rlp: &[u8]) -> Result<Signature, LedgerError>

Signs an Ethereum transaction’s RLP bytes (requires confirmation on the ledger).

Note that this does not apply EIP-155.

async fn sign_typed_data_( &self, hash_struct: &B256, domain: &Eip712Domain, ) -> Result<Signature, LedgerError>

Available on crate feature eip712 only.

async fn sign_payload( &self, command: INS, payload: &[u8], ) -> Result<Signature, LedgerError>

Helper function for signing either transaction data, personal messages or EIP712 derived structs.

fn path_to_bytes(derivation: &DerivationType) -> Vec<u8>